iOS/AddVPNScript/requirements.txt
facebook-wda==1.4.6
packaging~=21.3